+++ lib/distribute_reads/appropriate_pool.rb
@@ -3,30 +3,30 @@ module DistributeReads
     def _appropriate_pool(*args)
       if Thread.current[:distribute_reads]
         if Thread.current[:distribute_reads][:replica]
-          if @slave_pool.completely_blacklisted?
+          if @replica_pool.completely_blacklisted?
             raise DistributeReads::NoReplicasAvailable, "No replicas available" if Thread.current[:distribute_reads][:failover] == false
-            DistributeReads.log "No replicas available. Falling back to master pool."
-            @master_pool
+            DistributeReads.log "No replicas available. Falling back to primary pool."
+            @primary_pool
           else
-            @slave_pool
+            @replica_pool
           end
-        elsif Thread.current[:distribute_reads][:primary] || needs_master?(*args) || (blacklisted = @slave_pool.completely_blacklisted?)
+        elsif Thread.current[:distribute_reads][:primary] || needs_primary?(*args) || (blacklisted = @replica_pool.completely_blacklisted?)
           if blacklisted
             if Thread.current[:distribute_reads][:failover] == false
               raise DistributeReads::NoReplicasAvailable, "No replicas available"
             else
-              DistributeReads.log "No replicas available. Falling back to master pool."
+              DistributeReads.log "No replicas available. Falling back to primary pool."
             end
           end
-          stick_to_master(*args) if DistributeReads.by_default
-          @master_pool
+          stick_to_primary(*args) if DistributeReads.by_default
+          @primary_pool
         elsif in_transaction?
-          @master_pool
+          @primary_pool
         else
-          @slave_pool
+          @replica_pool
         end
       elsif !DistributeReads.by_default
-        @master_pool
+        @primary_pool
       else
         super
       end