+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.4.0 (unreleased)
+
+- Added support for Makara 0.6 and dropped support for Makara < 0.6
+- Made `distribute_reads` method private to behave like `Kernel` methods
+- Dropped support for Active Record < 5.2 and Ruby < 2.6
+
 ## 0.3.4 (2021-08-22)
 
 - Added support for Aurora Postgres replication lag